[ci] remove release step + rename to "CI / Tests"
The release was never actually used in rom-rb and I'll be adding a separate
workflow for the release part because now it is possible to request a
workflow from another workflow via workflow_dispatch trigger. We can also
push releases from individual repos now because it's finally possible to
set organization secrets for all repos.